api/base/testing: add common FacadeCaller patcher
And use it in usermanager, uniter and provisioner.
