+++ api/provisioner/export_test.go
@@ -4,25 +4,12 @@
 package provisioner
 
 import (
-	"github.com/juju/juju/api/base"
+	"github.com/juju/juju/api/base/testing"
 )
 
 // PatchFacadeCall patches the State's facade such that
 // FacadeCall method calls are diverted to the provided
 // function.
-func PatchFacadeCall(p patcher, st *State, f func(request string, params, response interface{}) error) {
-	p.PatchValue(&st.facade, &facadeWrapper{st.facade, f})
-}
-
-type patcher interface {
-	PatchValue(dest, value interface{})
-}
-
-type facadeWrapper struct {
-	base.FacadeCaller
-	facadeCall func(request string, params, response interface{}) error
-}
-
-func (f *facadeWrapper) FacadeCall(request string, params, response interface{}) error {
-	return f.facadeCall(request, params, response)
+func PatchFacadeCall(p testing.Patcher, st *State, f func(request string, params, response interface{}) error) {
+	testing.PatchFacadeCall(p, &st.facade, f)
 }
