+++ api/uniter/export_test.go
@@ -2,7 +2,9 @@
 // Licensed under the AGPLv3, see LICENCE file for details.
 package uniter
 
-import "github.com/juju/juju/api/base"
+import (
+	"github.com/juju/juju/api/base/testing"
+)
 
 var NewSettings = newSettings
 
@@ -12,26 +14,8 @@ var NewSettings = newSettings
 // It can also return an error to have the FacadeCall return an error.
 // The function returned by PatchResponses is a cleanup function that returns
 // the client to its original state.
-func PatchUnitResponse(u *Unit, responseFunc func(interface{}) error) func() {
-	orig := u.st.facade
-	u.st.facade = &resultCaller{responseFunc}
-	return func() {
-		u.st.facade = orig
-	}
-}
-
-type resultCaller struct {
-	setResult func(interface{}) error
-}
-
-func (f *resultCaller) FacadeCall(request string, params, response interface{}) error {
-	return f.setResult(response)
-}
-
-func (f *resultCaller) BestAPIVersion() int {
-	return 0
-}
-
-func (f *resultCaller) RawAPICaller() base.APICaller {
-	return nil
+func PatchUnitResponse(p testing.Patcher, u *Unit, responseFunc func(interface{}) error) {
+	testing.PatchFacadeCall(p, &u.st.facade, func(request string, params, response interface{}) error {
+		return responseFunc(response)
+	})
 }
